// nautilus-ionic/src/app/api/model/associationDTO.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InventoryBoxDTO } from './inventoryBoxDTO';


export interface AssociationDTO { 
    id?: number;
    name?: string;
    address?: string;
    email?: string;
    phone?: string;
    website?: string;
    remarks?: string;
    img64?: Blob;
    inventoryBoxes?: Array<InventoryBoxDTO>;
}